Refactor resource loading
